requirements.txt
alembic==1.6.5
coverage==6.2
coveralls==3.3.1
Flask==2.0.2
Flask-Classy==0.6.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
flask-restful==0.3.9
flask-swagger-ui==3.36.0
gunicorn==20.1.0
SQLAlchemy==1.4.29
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mysql-connector-python==8.0.27
python-dotenv==0.19.2
pylint==2.12.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
Werkzeug==2.0.2

